Solving merge bugs
