+++ app/src/androidTestEreferrals/java/org/eyeseetea/malariacare/common/configurationimporter/ConfigurationImporterUtil.java
@@ -1,42 +1,44 @@
 package org.eyeseetea.malariacare.common.configurationimporter;
 
 
-import android.content.Context;
-
 import com.raizlabs.android.dbflow.sql.language.Select;
 
+import org.eyeseetea.malariacare.data.database.model.AnswerDB;
 import org.eyeseetea.malariacare.data.database.model.CountryVersionDB;
 import org.eyeseetea.malariacare.data.database.model.HeaderDB;
+import org.eyeseetea.malariacare.data.database.model.MatchDB;
+import org.eyeseetea.malariacare.data.database.model.OptionAttributeDB;
 import org.eyeseetea.malariacare.data.database.model.OptionDB;
+import org.eyeseetea.malariacare.data.database.model.PhoneFormatDB;
 import org.eyeseetea.malariacare.data.database.model.ProgramDB;
 import org.eyeseetea.malariacare.data.database.model.QuestionDB;
 import org.eyeseetea.malariacare.data.database.model.QuestionOptionDB;
+import org.eyeseetea.malariacare.data.database.model.QuestionRelationDB;
+import org.eyeseetea.malariacare.data.database.model.QuestionThresholdDB;
+import org.eyeseetea.malariacare.data.database.model.SurveyDB;
 import org.eyeseetea.malariacare.data.database.model.TabDB;
-import org.eyeseetea.malariacare.data.database.utils.populatedb.UpdateDB;
 
-import java.io.IOException;
 import java.util.List;
 
 public class ConfigurationImporterUtil {
 
     public static void cleanUsedTables() {
-        HeaderDB.deleteAll();
-        TabDB.deleteAll();
-        ProgramDB.deleteAll();
         CountryVersionDB.deleteAll();
-
-
-        QuestionDB.deleteAll();
+        PhoneFormatDB.deleteAll();
+        SurveyDB.deleteAll();
+        ProgramDB.deleteAll();
+        TabDB.deleteAll();
+        HeaderDB.deleteAll();
+        AnswerDB.deleteAll();
+        OptionAttributeDB.deleteAll();
         OptionDB.deleteAll();
+        QuestionDB.deleteAll();
+        QuestionRelationDB.deleteAll();
+        MatchDB.deleteAll();
         QuestionOptionDB.deleteAll();
+        QuestionThresholdDB.deleteAll();
     }
 
-    public static void loadMetadataFromCSV(Context context) throws IOException {
-        UpdateDB.updatePrograms(context);
-        UpdateDB.updateTabs(context);
-        UpdateDB.updateHeaders(context);
-
-    }
 
     public static int getQuestionDBCount() {
         List<QuestionDB> questionDBS = QuestionDB.getAllQuestions();
@@ -55,5 +57,17 @@ public class ConfigurationImporterUtil {
         List<OptionDB> optionDBS = OptionDB.getAllOptions();
 
         return optionDBS.size();
+    }
+
+    public static int getProgramsDBCount() {
+        List<ProgramDB> programs = ProgramDB.getAllPrograms();
+
+        return programs.size();
+    }
+
+    public static int getPhoneFormatDBCount() {
+        List<PhoneFormatDB> formats = PhoneFormatDB.getAllPhoneFormats();
+
+        return formats.size();
     }
 }
